Split up main method & move method to fitting files
