+++ main.go
@@ -1,14 +1,10 @@
 package main
 
 import (
-	"errors"
 	"fmt"
 	"gopkg.in/alecthomas/kingpin.v2"
-	"io/ioutil"
 	"os"
-	"os/exec"
 	"path"
-	"strings"
 )
 
 var (
@@ -20,32 +16,6 @@ var (
 	hostname        = os.Getenv("HOST")
 )
 
-func getHostConfig(hostname string, config Config) (Host, error) {
-	for _, entry := range config.Hosts {
-		if entry.Name == hostname {
-			return entry, nil
-		}
-	}
-	return Host{}, errors.New("file for hostname not found")
-}
-
-func ReadPackagesFromFile(fileName string, baseDir string) ([]string, error) {
-	filePath := fileName
-	if baseDir != "" {
-		filePath = baseDir + "/" + fileName
-	}
-	data, fileErr := ioutil.ReadFile(filePath)
-	if fileErr != nil {
-		return []string{}, fileErr
-	}
-
-	text := string(data[:])
-
-	packages := strings.Split(text, "\n")
-
-	return packages, nil
-}
-
 func main() {
 	res := kingpin.MustParse(app.Parse(os.Args[1:]))
 	config := ParseConfigFile(*configFile)
@@ -55,7 +25,7 @@ func main() {
 	}
 }
 func sync(config Config, baseDir string) {
-	hostConfig, err := getHostConfig(hostname, config)
+	hostConfig, err := GetHostConfig(hostname, config)
 	versionedPackages, err := ReadPackagesFromFile(hostConfig.File, baseDir)
 	systemPackages, err := GetInstalledPackages()
 	if err != nil {
@@ -68,73 +38,11 @@ func sync(config Config, baseDir string) {
 	userInput := UserInput{dryRun: *dryRunFlag, versionedPackages: versionedPackages, systemPackages: systemPackages, ignoredPackages: ignoredPackages}
 	userInput.HandleHostPackagesChange()
 	userInput.HandleSubscribedPackageChanges(hostConfig, config, baseDir)
-	commit(userInput, hostConfig, baseDir)
-}
-func commit(userInput UserInput, hostConfig Host, baseDir string) {
-	if !*dryRunFlag {
-		filePath := baseDir + "/" + hostConfig.File
-		systemPackages := []byte(strings.Join(userInput.systemPackages, "\n"))
-		err := ioutil.WriteFile(filePath, systemPackages, os.ModePerm)
-		if err != nil {
-			fmt.Println("Failed to save versioned packages:", err)
-		}
-		ignoredPackages := []byte(strings.Join(userInput.ignoredPackages, "\n"))
-		filePath = baseDir + "/" + hostConfig.IgnoreFile
-		err = ioutil.WriteFile(filePath, ignoredPackages, os.ModePerm)
-		if err != nil {
-			fmt.Println("Failed to save ignored packages:", err)
-		}
-	} else {
-		fmt.Println("New versioned packages:")
-		for _, v := range userInput.systemPackages {
-			fmt.Println(v)
-		}
-		fmt.Println("New ignored packages:")
-		for _, v := range userInput.ignoredPackages {
-			fmt.Println(v)
-		}
-	}
-}
-func filterComparisonResult(ignoredPackages []string, compareResult CompareResult) CompareResult {
-	toRemove := clearList(compareResult.Removed, ignoredPackages)
-	toAdd := clearList(compareResult.Added, ignoredPackages)
-	return CompareResult{toAdd, toRemove, compareResult.Unchanged}
+	CommitPackageChanges(userInput, hostConfig, baseDir)
 }
 
-func clearList(toClear []string, ignoredPackages []string) []string {
-	var result []string
-	for _, value := range toClear {
-		if contains(ignoredPackages, value) {
-			continue
-		} else {
-			result = append(result, value)
-		}
-	}
-	return result
-}
 
-func GetInstalledPackages() ([]string, error) {
-	var packages []string
-	if hostPackageFile != nil {
-		packages, _ = ReadPackagesFromFile((*hostPackageFile).Name(), "")
-	} else {
-		cmd := exec.Command("yaourt", "-Qqe")
-		result, err := cmd.Output()
-		if err != nil {
-			return []string{}, nil
-		}
-		resultAsString := string(result)
-		packages = strings.Split(resultAsString, "\n")
-	}
-	return clearPackages(packages), nil
-}
 
-func clearPackages(packages []string) []string {
-	var clearedPackages []string
-	for _, v := range packages {
-		if len(v) > 1 {
-			clearedPackages = append(clearedPackages, v)
-		}
-	}
-	return clearedPackages
-}
+
+
+
